Made some improvements to the logger.
